+++ CHANGES.rst
@@ -97,6 +97,9 @@ New Features
   - A new sub-package, `astropy.vo.samp`, is now available (this was previously
     the SAMPy package, which has been refactored for use in Astropy). [#1907]
 
+  - Enhanced functionalities for `~astropy.vo.client.vos_catalog.VOSCatalog`
+    and `~astropy.vo.client.vos_catalog.VOSDatabase`. [#1206]
+
 - ``astropy.wcs``
 
 API Changes